Improve performance of 'rs_loadsig_file'
The performance of rs_loadsig_file is bottlenecked by rs_loadsig_add_sum,
which uses realloc to increase the size of job->signature by one on each
invocation, which is slow.

This commit improves performance by increasing the size of job->signature
exponentially - in essence treating it as a buffer.  On Windows using
Mingw64 this improves the performance time of `rdiff delta` by 11x.